Split classes into separate files under kilauncher dir
