+++ kilauncher/defaults.py
@@ -1,0 +1,3 @@
+DEFAULT_LAUNCHER_SIZE = '240x80'
+DEFAULT_ICON_SIZE = '64x64'
+DEFAULT_STYLESHEET = '/etc/kilauncher/stylesheet.css'
